Move models.py into models folder.
Models.py files is getting ridicilously big. Put it into a folder, so at least future models can have separate file.

Refs HEL-168
