+++ hkm/management/commands/create_initial_page_content.py
@@ -1,8 +1,9 @@
 # -*- coding: utf-8 -*-
 from __future__ import unicode_literals
+
 from django.core.management.base import BaseCommand
 
-from hkm.models import PageContent
+from hkm.models.models import PageContent
 
 pages = [{
     "name": u"About",
