+++ hkm/tasks.py
@@ -8,7 +8,7 @@ from django.core.mail import send_mail
 from django.template.loader import render_to_string
 from django.utils.translation import ugettext as _
 
-from hkm.models import Feedback
+from hkm.models.models import Feedback
 
 LOG = logging.getLogger(__name__)
